Fix PiOS aarch64 cross build, enable drm-kms-egl on bookworm, add CI

+++ shell/backend/drm_kms_egl/driver_probe.cc
@@ -15,6 +15,7 @@
  */
 
 #include "backend/drm_kms_egl/driver_probe.h"
+#include "logging/logging.h"
 
 #include <drm_fourcc.h>
 #include <drm_mode.h>
@@ -60,7 +61,8 @@ std::string GetDriverName(const int drm_fd) {
   if (!v) {
     return {};
   }
-  std::string name(v->name ? v->name : "", v->name_len > 0 ? v->name_len : 0);
+  std::string name(v->name ? v->name : "",
+                   static_cast<size_t>(v->name_len > 0 ? v->name_len : 0));
   drmFreeVersion(v);
   return name;
 }
